Refactor ordinary_set: add tracing, fix borrow/move issues, clean warnings
