+++ core/engine/Cargo.toml
@@ -90,6 +90,7 @@ xsum = ["dep:xsum"]
 native-backtrace = []
 
 [dependencies]
+tracing = { version = "0.1", features = ["log"] }
 tag_ptr.workspace = true
 boa_interner.workspace = true
 boa_gc = { workspace = true, features = ["thin-vec", "boa_string", "arrayvec"] }
